+++ lib/Fapi/src/Batch/FapiSubscribeWebhooksBatch.ts
@@ -34,6 +34,7 @@ export default class FapiSubscribeWebhooksBatch extends ABatchNode {
         const subscribedWebhooks = await webhookRepository.findMany({
             users: [applicationInstall.getUser()],
             apps: [applicationInstall.getName()],
+            sdks: [applicationInstall.getSdk()],
         });
 
         const unsubscribedWebhook = webhookSubscriptions.filter(
@@ -71,6 +72,7 @@ export default class FapiSubscribeWebhooksBatch extends ABatchNode {
                 .setWebhookId(String(responseDto.getJsonBody().id))
                 .setApplication(applicationInstall.getName())
                 .setUser(applicationInstall.getUser())
+                .setSdk(applicationInstall.getSdk())
                 .setTopology(topology)
                 .setNode(node)
                 .setToken(token)