+++ lib/Pipedrive/src/PipedriveApplication.ts
@@ -86,7 +86,12 @@ export default class PipedriveApplication extends ABasicApplication implements I
         url: string,
     ): RequestDto {
         return this.getRequestDto(
-            ProcessDto.createForFormRequest(NAME, applicationInstall.getUser(), crypto.randomUUID()),
+            ProcessDto.createForFormRequest(
+                NAME,
+                applicationInstall.getUser(),
+                applicationInstall.getSdk(),
+                crypto.randomUUID(),
+            ),
             applicationInstall,
             HttpMethods.POST,
             `${PIPEDRIVE_URL}/v1/webhooks`,
@@ -123,7 +128,12 @@ export default class PipedriveApplication extends ABasicApplication implements I
         webhook: Webhook,
     ): RequestDto {
         return this.getRequestDto(
-            ProcessDto.createForFormRequest(NAME, applicationInstall.getUser(), crypto.randomUUID()),
+            ProcessDto.createForFormRequest(
+                NAME,
+                applicationInstall.getUser(),
+                applicationInstall.getSdk(),
+                crypto.randomUUID(),
+            ),
             applicationInstall,
             HttpMethods.DELETE,
             `${PIPEDRIVE_URL}/v1/webhooks/${webhook.getWebhookId()}`,