+++ lib/Shipstation/src/ShipstationApplication.ts
@@ -81,7 +81,12 @@ export default class ShipstationApplication extends ABasicApplication implements
         subscription: WebhookSubscription,
         url: string,
     ): RequestDto {
-        const request = ProcessDto.createForFormRequest(NAME, applicationInstall.getUser(), crypto.randomUUID());
+        const request = ProcessDto.createForFormRequest(
+            NAME,
+            applicationInstall.getUser(),
+            applicationInstall.getSdk(),
+            crypto.randomUUID(),
+        );
         return this.getRequestDto(
             request,
             applicationInstall,
@@ -105,7 +110,12 @@ export default class ShipstationApplication extends ABasicApplication implements
     }
 
     public getWebhookUnsubscribeRequestDto(applicationInstall: ApplicationInstall, webhook: Webhook): RequestDto {
-        const request = ProcessDto.createForFormRequest(NAME, applicationInstall.getUser(), crypto.randomUUID());
+        const request = ProcessDto.createForFormRequest(
+            NAME,
+            applicationInstall.getUser(),
+            applicationInstall.getSdk(),
+            crypto.randomUUID(),
+        );
         return this.getRequestDto(
             request,
             applicationInstall,