+++ lib/Google/Common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @orchesty/connector-google-common
 
+## 2.0.0
+
+### Major Changes
+
+- Use @orchesty/nodejs-sdk ^5.0
+
 ## 1.0.2
 
 ### Patch Changes