+++ lib/Xero/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @orchesty/connector-xero
 
+## 2.0.0
+
+### Major Changes
+
+- Use @orchesty/nodejs-sdk ^5.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @orchesty/connector-digitoo@2.0.0
+
 ## 1.0.1
 
 ### Patch Changes